requirements.txt
django==2.0
djangorestframework==3.7.3
ipython==6.2.1
psycopg2==2.7.3
python-dotenv==0.7.1
django-extensions==1.9.8
ipdb==0.10.3
